+++ UnitTest/Suporting/test_csvWriterLogger.py
@@ -1,3 +1,0 @@
-import pytest
-from GabesPythonToolBox.Suporting import *
-#TODO