+++ Utility/ConfigManager.py
@@ -2,7 +2,6 @@ import json
 import yaml
 import os
 import copy
-
 from ..Utility.Debug import *
 from ..DataProsesing.DataRW import *